pyproject.toml
[project]
name = "broai-research-rocket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "broai==0.1.13",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "torch==2.6.0",
    "uvicorn>=0.34.2",
]
